Cleaning up Bucket Shadowing
There were some questions on clarity of bucket shadowing unrelated to
documentation, but overall the feedback was just that it needed to be
made clear there were two components, one was the CBS app bucket, one
was the SG dedicated bucket, and that they "bucket shadow" each other.
Needs review.