Replaced Google Analytics with Google Tag Manager.
Change being made per request of Kari Hodgson on April 14, relayed through
Ali and Wayne.

Note that I tried to find a way to include this through the site.xml, but
could not find a way to escape out the & included in the supplied script.
Looking around, the pattern seemed to be to use a CDATA section in the
html.xslt, so I followed the same approach which worked.