Replaced Google Analytics with Google Tag Manager.
Per discussion with Kari, Wayne, Ali and myself, we've decided to put this
back into place and monitor.  The revert was done despite not really
knowing if there was an effect.  We want this in place for upcoming
events.

Note that I tried to find a way to include this through the site.xml, but
could not find a way to escape out the & included in the supplied script.
Looking around, the pattern seemed to be to use a CDATA section in the
html.xslt, so I followed the same approach which worked.