Rewrote paragraph introducing views.
The paragraph that was there seemed to assume Couchbase Server knowledge
which may have been the case for the original author but isn't right
for the audience we expect here.  I've rewritten it to make it a bit
more clear what views are rather than try to compare them to something
the user probably doesn't know.