To have synthax highlighting the code tag should have a css class of language-type where type can be swift, java, c.
The code-block template is responsible for the rendering of the snippet but
the language type is computed as a variable in the code-set template.

That's why I pass it as a param in the apply-templates call (feature available in xslt 2
when enabling the tunnel flag).

Could refactor this in the future but it works for now.

Fix #251